+++ SupermarketApp/SupermarketApp/ViewModel/CashierVM.cs
@@ -312,7 +312,7 @@ namespace SupermarketApp.ViewModel
 
                             _stocksBLL.UpdateStockQuantity(SearchedStock, SearchedStock.Quantity + item.Quantity);
                             Receipt.Total -= item.Subtotal;
-                            Receipt.Total = Math.Round(Receipt.Total, 2);
+                            Receipt.Total = (float)Math.Round(Receipt.Total, 2);
                             ProductsList.Remove(item);
                         }
                         else
@@ -320,9 +320,9 @@ namespace SupermarketApp.ViewModel
                             _stocksBLL.UpdateStockQuantity(SearchedStock, SearchedStock.Quantity - quantity);
                             Receipt.Total -= item.Subtotal;
                             item.Quantity += quantity;
-                            item.Subtotal = Math.Round(item.Stock.SalePrice * item.Quantity, 2);
+                            item.Subtotal = (float)Math.Round(item.Stock.SalePrice * item.Quantity, 2);
                             Receipt.Total += item.Subtotal;
-                            Receipt.Total = Math.Round(Receipt.Total, 2);
+                            Receipt.Total = (float)Math.Round(Receipt.Total, 2);
                             if (SearchedStock.Quantity == 0)
                                 SearchedStock = null;
                         }
@@ -342,9 +342,9 @@ namespace SupermarketApp.ViewModel
                 stockReceipt.Receipt = Receipt;
                 stockReceipt.Stock = SearchedStock;
                 stockReceipt.Quantity = quantity;
-                stockReceipt.Subtotal = Math.Round(SearchedStock.SalePrice * stockReceipt.Quantity, 2);
+                stockReceipt.Subtotal = (float)Math.Round(SearchedStock.SalePrice * stockReceipt.Quantity, 2);
                 Receipt.Total += stockReceipt.Subtotal;
-                Receipt.Total = Math.Round(Receipt.Total, 2);
+                Receipt.Total = (float)Math.Round(Receipt.Total, 2);
                 ProductsList.Add(stockReceipt);
                 if (SearchedStock.Quantity == 0)
                     SearchedStock = null;
